+++ src/components/Input/TextInput.vue
@@ -131,6 +131,10 @@ export default {
       type: Boolean,
       default: false,
     },
+    showClear: {
+      type: Boolean,
+      default: false,
+    }
   },
   emits: ['icon-left-click', 'icon-right-click', 'clear'],
   data() {
@@ -218,7 +222,6 @@ export default {
     top: 50%;
     transform: translateY(-50%);
     right: $unnnic-space-4;
-
     display: flex;
     align-items: center;
     gap: $unnnic-space-2;